Separate Localizer and PromptManager
